[core] Move parameterized Entry API tests into separate file
